+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Latest
 
+# v5.1.0
+
+* Introduces backwards compatibility with Conjur 4.x for most API methods.
+* Adds the configuration setting `version`, which is auto-populated from the environment variable `CONJUR_VERSION`.
+
 # v5.0.0
 
 * Provides compatibility with [cyberark/conjur](https://github.com/cyberark/conjur), Conjur 5 CE.